requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
django-filter==21.1
drf-yasg==1.21.3
boto3~=1.24.27
django-rest-swagger==2.2.0
